+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All noteable changes will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- **New Features**:
+     - Automod System:
+          - 1 new command: (`/automod_setup`)
+          - 3 new sub-commands: (`add`, `remove`, `toggle`)
+               - Read up on the new command **[here](./docs/COMMAND_GUIDE_README/GUILD_N_UTILITY_COMMANDS.md#automod-setup)**
+
 ## [3.0.0] - [2026-02-16]
 ### Added
 - **New Commands**: